test/regression/reftest/vendor/blink/fast/dom/XMLHttpRequest-legacy-event-listener.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/wpt-test-pre.js"></script>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description("Test addEventListener() and removeEventListener() fail silently if arguments are missing.");

shouldThrow('new XMLHttpRequest().addEventListener("foo")');
shouldThrow('new XMLHttpRequest().removeEventListener("bar")');

shouldThrow('new XMLHttpRequest().addEventListener()');
shouldThrow('new XMLHttpRequest().removeEventListener()');
</script>
<script src="../../resources/js-test-post.js"></script>
<script src="../../resources/wpt-test-post.js"></script>
</body>
</html>
